refactor: Adapter 계층(web, persistence)의 클래스 접근 제어자를 package-private으로 변경하여, 상위 계층(application 등)에서 어댑터 구현체를 직접 호출하지 못하도록 방지
#33
