+++ src/main/java/com/yoger/productserviceorganization/product/adapters/messaging/kafka/consumer/EventDeduplicateService.java
@@ -1,0 +1,7 @@
+package com.yoger.productserviceorganization.product.adapters.messaging.kafka.consumer;
+
+interface EventDeduplicateService {
+    boolean isDuplicate(String eventId);
+
+    void putKey(String eventId);
+}
