--- src/main/java/com/yoger/productserviceorganization/product/adapters/messaging/kafka/consumer/dedup/EventDeduplicateService.java
@@ -1,7 +1,0 @@
-package com.yoger.productserviceorganization.product.adapters.messaging.kafka.consumer.dedup;
-
-public interface EventDeduplicateService {
-    boolean isDuplicate(String eventId);
-
-    void putKey(String eventId);
-}
