+++ src/main/java/com/yoger/productserviceorganization/product/adapters/persistence/jpa/JpaOutboxRepository.java
@@ -2,5 +2,5 @@ package com.yoger.productserviceorganization.product.adapters.persistence.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaOutboxRepository extends JpaRepository<OutboxEventJpaEntity, String> {
+interface JpaOutboxRepository extends JpaRepository<OutboxEventJpaEntity, String> {
 }
