+++ src/main/java/com/yoger/productserviceorganization/product/adapters/persistence/jpa/OutboxPersistenceAdapter.java
@@ -1,17 +1,15 @@
-package com.yoger.productserviceorganization.product.adapters.persistence;
+package com.yoger.productserviceorganization.product.adapters.persistence.jpa;
 
-import com.yoger.productserviceorganization.product.adapters.persistence.jpa.JpaOutboxRepository;
 import com.yoger.productserviceorganization.product.application.port.out.ClearOutboxEventPort;
 import com.yoger.productserviceorganization.product.application.port.out.SaveOutboxEventPort;
 import com.yoger.productserviceorganization.product.domain.event.OutboxEvent;
-import com.yoger.productserviceorganization.product.mapper.OutboxEventMapper;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 @Repository
 @RequiredArgsConstructor
-public class OutboxPersistenceAdapter implements SaveOutboxEventPort, ClearOutboxEventPort {
+class OutboxPersistenceAdapter implements SaveOutboxEventPort, ClearOutboxEventPort {
     private final JpaOutboxRepository jpaOutboxRepository;
 
     @Override
