+++ src/main/java/com/yoger/productserviceorganization/product/adapters/persistence/jpa/ProductJpaEntity.java
@@ -27,7 +27,7 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProductJpaEntity {
+class ProductJpaEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
