+++ src/main/java/com/yoger/productserviceorganization/product/adapters/web/controller/ProductController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/products")
 @RequiredArgsConstructor
-public class ProductController {
+class ProductController {
     private final GetProductQuery getProductQuery;
 
     @GetMapping
